+++ packages/stacked_services/example/lib/app/locator.dart
@@ -1,8 +1,0 @@
-import 'package:get_it/get_it.dart';
-import 'package:injectable/injectable.dart';
-import 'locator.config.dart';
-
-final locator = GetIt.instance;
-
-@injectableInit
-void setupLocator() => $initGetIt(locator);