+++ packages/stacked_tools/example/lib/enums/bottom_sheet_type.dart
@@ -1,0 +1,4 @@
+enum BottomSheetType {
+  notice,
+}
+