+++ packages/stacked_tools/example/lib/enums/dialog_type.dart
@@ -1,0 +1,4 @@
+enum DialogType {
+  infoAlert,
+}
+