+++ packages/stacked_tools/lib/src/mixins/project_structure_validator_mixin.dart
@@ -1,6 +1,6 @@
+import 'package:stacked_tools/src/constants/message_constants.dart';
 import 'package:stacked_tools/src/exceptions/invalid_stacked_structure_exception.dart';
 import 'package:stacked_tools/src/locator.dart';
-import 'package:stacked_tools/src/message_constants.dart';
 import 'package:stacked_tools/src/services/file_service.dart';
 
 mixin ProjectStructureValidator {
@@ -13,7 +13,7 @@ mixin ProjectStructureValidator {
       throw InvalidStackedStructureException(kInvalidRootDirectory);
     }
 
-    if (!(await _fileService.isStakedApplication(outputPath: outputPath))) {
+    if (!(await _fileService.isStackedApplication(outputPath: outputPath))) {
       throw InvalidStackedStructureException(kInvalidStackedStructure);
     }
   }