+++ packages/stacked_tools/playground/lib/app/app.dart
@@ -1,9 +1,16 @@
-import 'package:example/ui/views/order_details/order_details_view.dart';
 import 'package:stacked/stacked_annotations.dart';
+import 'package:playground/ui/views/home/home_view.dart';
+import 'package:playground/ui/views/startup/startup_view.dart';
+import 'package:stacked_services/stacked_services.dart';
 // @stacked-import
 
 @StackedApp(routes: [
-  MaterialRoute(page: OrderDetailsView),
-// @stacked-route
+  MaterialRoute(page: StartupView),
+  MaterialRoute(page: HomeView),
+  // @stacked-route
+], dependencies: [
+  LazySingleton(classType: NavigationService),
+  LazySingleton(classType: DialogService),
+  LazySingleton(classType: BottomSheetService),
 ])
 class App {}